+++ lib/qsu.dart
@@ -2,3 +2,4 @@ library;
 
 export 'src/format.dart';
 export 'src/string.dart';
+export 'src/verify.dart';