+++ lib/src/verify.dart
@@ -1,0 +1,17 @@
+bool contains(dynamic str, dynamic search, {bool exact = false}) {
+  if (search.runtimeType == String) {
+    return str.length < 1 ? false : str.indexOf(search) != -1;
+  }
+
+  for (int i = 0, iLen = search.length; i < iLen; i += 1) {
+    if (exact) {
+      if (str == search[i]) {
+        return true;
+      }
+    } else if (str.indexOf(search[i]) != -1) {
+      return true;
+    }
+  }
+
+  return false;
+}